Move Commands into it's own file
