+++ src/inspector/tools/hotreloader.cpp
@@ -7,6 +7,7 @@
 #include <QDir>
 #include <QFile>
 #include <QFileInfo>
+#include <QJsonObject>
 #include <QQmlApplicationEngine>
 #include <QStandardPaths>
 #include <QWindow>
@@ -17,6 +18,10 @@
 #include "networkrequest.h"
 #include "qmlengineholder.h"
 #include "tasks/function/taskfunction.h"
+#include "command.h"
+#include "commandhandler.h"
+
+#include "../inspector.h"
 
 namespace {
 Logger logger("QMLHotReload");
@@ -43,6 +48,8 @@ Hotreloader::Hotreloader(QObject* parent, QQmlEngine* target)
       dir.removeRecursively();
     }
   });
+
+  registerDevCommands();
 }
 
 void Hotreloader::annonceReplacedFile(const QUrl& path) {
@@ -168,6 +175,36 @@ QUrl Hotreloader::HotReloadInterceptor::intercept(
   }
   return url;
 }
+
+void Hotreloader::registerDevCommands(){
+  Inspector* i = qobject_cast<Inspector*>(parent());
+  if (!i) {
+    return;
+  }
+
+
+  i->registerCommand(
+   InspectorTools::Command{"live_reload", "Live reload file X", 1,
+                          [this](const QList<QByteArray>& args) {
+                            auto url = QUrl(args.at(1));
+                            annonceReplacedFile(url);
+                            return QJsonObject();
+                          }});
+   i->registerCommand(
+      InspectorTools::Command{"reload_window", "Reload the whole window", 0,
+                              [this](const QList<QByteArray>& args) {
+                                reloadWindow();
+                                return QJsonObject();
+                              }});
+
+     i->registerCommand(InspectorTools::Command{
+       "reset_live_reload", "Reset all hot reloaded files", 0,
+       [this](const QList<QByteArray>& args) {
+        resetAllFiles();
+         return QJsonObject();
+       }});
+}
+
 
 
 }  // namespace InspectorTools
